refactor(usePlaceOrder): extract all form-related logic into the `usePlaceOrder` hook
